Refactor PriceConvertion complete
