+++ src/components/CoinDetails/PriceConvertion/ExchangeAmount.js
@@ -2,7 +2,11 @@ import PropTypes from 'prop-types';
 import { cleanValue, handleValue } from '../../../hooks/helperFuncs';
 import CurrencyFilter from './CurrencyFilter';
 
-const ExchangeAmount = ({ exchangeAmount, handleExchangeAmount }) => (
+const ExchangeAmount = ({
+  exchangeAmount,
+  handleExchangeAmount,
+  handleExchangeRate,
+}) => (
   <label
     htmlFor="exAmount"
     className="filter-label"
@@ -14,7 +18,7 @@ const ExchangeAmount = ({ exchangeAmount, handleExchangeAmount }) => (
       value={cleanValue(exchangeAmount)}
       onInput={(e) => handleValue(e, handleExchangeAmount)}
     />
-    <CurrencyFilter />
+    <CurrencyFilter handleExchangeRate={handleExchangeRate} />
   </label>
 );
 
@@ -23,4 +27,5 @@ export default ExchangeAmount;
 ExchangeAmount.propTypes = {
   exchangeAmount: PropTypes.number.isRequired,
   handleExchangeAmount: PropTypes.func.isRequired,
+  handleExchangeRate: PropTypes.func.isRequired,
 };
