+++ src/redux/coin-details/coinDetailsSlice.js
@@ -37,22 +37,19 @@ const coinDetailsSlice = createSlice({
         ...state,
         isLoading: true,
       }))
-      .addCase(fetchCoinDetails.fulfilled, (state, { payload }) => {
-        console.log(payload.image.small, 'slice');
-        return ({
-          ...state,
-          isLoading: false,
-          coinDetails: {
-            name: payload.name,
-            symbol: payload.symbol,
-            image: payload.image.small,
-            currentPrice: payload.market_data.current_price,
-            marketData: payload.market_data,
-            description: payload.description.en,
-          },
-          ready: true,
-        });
-      })
+      .addCase(fetchCoinDetails.fulfilled, (state, { payload }) => ({
+        ...state,
+        isLoading: false,
+        coinDetails: {
+          name: payload.name,
+          symbol: payload.symbol,
+          image: payload.image.small,
+          currentPrice: payload.market_data.current_price,
+          marketData: payload.market_data,
+          description: payload.description.en,
+        },
+        ready: true,
+      }))
       .addCase(fetchCoinDetails.rejected, (state, action) => ({
         ...state,
         isLoading: false,
